Add support for more candid types (#26)
This adds `arg` support for records (of strings & records).

The arg inference is changed a bit and tested against agent-go's
`didc`-style encoding of strings (e.g. strings like `"record {...}"`).

A couple of helper functions are added to the provider. These are used
to _encode_ (`did_encode`, used mostly in testing and used implicitely
when using `arg`) and _annotate_ values with types (`did_text`,
`did_record`) to avoid ambiguity (i.e. for soon-to-come principal
support).